4. Тесты для стороны прямоугольника. [#159460]

+++ src/main/java/ru/job4j/condition/RectangleArea.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class RectangleArea {
-    public static double square(double p, double k) {
+    public static double square(int p, double k) {
       double h = p / (2 * (k + 1));
       double l = h * k;
       double result = l * h;
